chore: let only application recreate the table

+++ app/config/postgres.py
@@ -35,8 +35,9 @@ class ResourcesManager:
                 )
                 async with engine.begin() as conn:
                     await conn.execute(text('CREATE EXTENSION IF NOT EXISTS "uuid-ossp"'))
-                    await conn.run_sync(Base.metadata.drop_all)
-                    await conn.run_sync(Base.metadata.create_all)
+                    if settings.RECREATE_POSTGRES_TABLES:
+                        await conn.run_sync(Base.metadata.drop_all)
+                        await conn.run_sync(Base.metadata.create_all)
             except IntegrityError as e:
                 #  "uuid-ossp" already created
                 #  "CREATE EXTENSION IF NOT EXISTS" is not concurrently safe
